requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
Pillow>=8.0.0
matplotlib>=3.3.0
numpy>=1.19.0
requests>=2.25.0
